+++ Cargo.toml
@@ -109,6 +109,7 @@ members = [
     "sdk/clock",
     "sdk/decode-error",
     "sdk/derivation-path",
+    "sdk/epoch-schedule",
     "sdk/feature-set",
     "sdk/gen-headers",
     "sdk/hash",
@@ -400,6 +401,7 @@ solana-define-syscall = { path = "define-syscall", version = "=2.1.0" }
 solana-derivation-path = { path = "sdk/derivation-path", version = "=2.1.0" }
 solana-download-utils = { path = "download-utils", version = "=2.1.0" }
 solana-entry = { path = "entry", version = "=2.1.0" }
+solana-epoch-schedule = { path = "sdk/epoch-schedule", version = "=2.1.0" }
 solana-faucet = { path = "faucet", version = "=2.1.0" }
 solana-feature-set = { path = "sdk/feature-set", version = "=2.1.0" }
 solana-fee = { path = "fee", version = "=2.1.0" }
